requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dateutil==2.8.2
PyPDF2==3.0.1
mysql-connector-python==8.1.0
requests==2.31.0
pydantic==2.0.0
pydantic-settings==2.0.0
Werkzeug==2.3.7
